AU: use the provisional doctor-certified deaths
The previous version was comparing two datasets that can't accurately be
compared. The pre-2020 dataset contained every death registered, but the 2020
dataset only contained doctor-certified deaths, which is why the 2020 counts
were comparatively low.

Doctor-certified deaths are registered within days or weeks, while
coroner-certified deaths can be delayed for weeks or months.

This change makes it so that we're using only doctor-certified death datasets.
When the death data for 2020 is complete, then this might be worth revisiting
so that we can get a picture of all deaths occurred.